추가로 푼 3문제

+++ 기초 100제/c++/1030.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 int main(){
-    unsigned int a;
+    long long int a;
     std::cin>>a;
     std::cout<<a;
 }
